refactor: 妖精のsubsystemからspeech部分を別のsubsystemとして切り出す
