+++ src/main/scala/com/github/unchama/seichiassist/subsystems/vote/subsystems/fairyspeech/bukkit/BukkitFairySpeechGateway.scala
@@ -1,8 +1,8 @@
-package com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.bukkit.gateway
+package com.github.unchama.seichiassist.subsystems.vote.subsystems.fairyspeech.bukkit
 
 import cats.effect.Sync
 import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.domain.property.FairyMessage
-import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairy.domain.speech.FairySpeechGateway
+import com.github.unchama.seichiassist.subsystems.vote.subsystems.fairyspeech.domain.FairySpeechGateway
 import com.github.unchama.targetedeffect.commandsender.MessageEffectF
 import com.github.unchama.targetedeffect.player.FocusedSoundEffectF
 import org.bukkit.ChatColor.{AQUA, BOLD, RESET}
